Rewrote command interpretation algorithm, fixing bugs related to arguments that are also filenames
Update filesDB (pyo -u) before use.

After reading in the user command from dmenu, matching algorithm now (i) tries to find an as long as possible entry also in filesDB and (ii) uses the remainder of the line as arguments. This is cleaner and less error-prone.
This assumes the first word given is part of the command, but to my knowledge this must always be the case for bash syntax. It also requires the user to supply the full path to any files given in the argument, just like a normal bash command.